requirements-prd.txt
minio
loguru
pydantic
pyyaml